requirements.txt
Flask
Flask-Cors
flask-socketio
PyJWT
numpy
rich
jsonpickle
apscheduler
